# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AuthFailureReason": {
				"enum": [
					"Wrong Password",
					"Unknown Cause"
				],
				"type": "string"
			},
			"FailedAuthResponse": {
				"properties": {
					"reason": {
						"$ref": "#/components/schemas/AuthFailureReason"
					}
				},
				"required": [
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessfulAuthResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HTTPResponse_FailedAuthResponse.SuccessfulAuthResponse_": {
				"properties": {
					"error": {
						"$ref": "#/components/schemas/FailedAuthResponse"
					},
					"success": {
						"$ref": "#/components/schemas/SuccessfulAuthResponse"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserParams": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginUserParams": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeniedPasswordResetResponseReason": {
				"enum": [
					"Too Many Attempts"
				],
				"type": "string"
			},
			"DeniedPasswordResetResponse": {
				"properties": {
					"reason": {
						"$ref": "#/components/schemas/DeniedPasswordResetResponseReason"
					}
				},
				"required": [
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessfulPasswordResetResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"HTTPResponse_DeniedPasswordResetResponse.SuccessfulPasswordResetResponse_": {
				"properties": {
					"error": {
						"$ref": "#/components/schemas/DeniedPasswordResetResponse"
					},
					"success": {
						"$ref": "#/components/schemas/SuccessfulPasswordResetResponse"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"RequestPasswordResetParams": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FailedToCreatePostResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"SuccessfulPostCreationResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"HTTPResponse_FailedToCreatePostResponse.SuccessfulPostCreationResponse_": {
				"properties": {
					"error": {
						"$ref": "#/components/schemas/FailedToCreatePostResponse"
					},
					"success": {
						"$ref": "#/components/schemas/SuccessfulPostCreationResponse"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PostPrivacySetting": {
				"enum": [
					"Tier2AndTier3"
				],
				"type": "string"
			},
			"PostDurationSetting": {
				"enum": [
					"Forever"
				],
				"type": "string"
			},
			"CreatePostParams": {
				"properties": {
					"imageId": {
						"type": "string"
					},
					"caption": {
						"type": "string"
					},
					"visibility": {
						"$ref": "#/components/schemas/PostPrivacySetting"
					},
					"duration": {
						"$ref": "#/components/schemas/PostDurationSetting"
					},
					"title": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"collaboratorUsernames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"scheduledPublicationTimestamp": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"imageId",
					"caption",
					"visibility",
					"duration",
					"title",
					"price",
					"collaboratorUsernames",
					"scheduledPublicationTimestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SecuredHTTPRequest_CreatePostParams_": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CreatePostParams"
					}
				},
				"required": [
					"accessToken",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FailedToUpdateUserSettingsResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"SuccessfulUpdateToUserSettingsResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"HTTPResponse_FailedToUpdateUserSettingsResponse.SuccessfulUpdateToUserSettingsResponse_": {
				"properties": {
					"error": {
						"$ref": "#/components/schemas/FailedToUpdateUserSettingsResponse"
					},
					"success": {
						"$ref": "#/components/schemas/SuccessfulUpdateToUserSettingsResponse"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DefaultPostPrivacySetting": {
				"enum": [
					"PublicAndGuestCheckout"
				],
				"type": "string"
			},
			"SetUserSettingsParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"profileVisibility": {
						"$ref": "#/components/schemas/DefaultPostPrivacySetting"
					},
					"bannedUsernames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"username",
					"bio",
					"website",
					"profileVisibility",
					"bannedUsernames"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SecuredHTTPRequest_SetUserSettingsParams_": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/SetUserSettingsParams"
					}
				},
				"required": [
					"accessToken",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeniedGetUserPageResponseReason": {
				"enum": [
					"Blocked"
				],
				"type": "string"
			},
			"DeniedGetUserPageResponse": {
				"properties": {
					"reason": {
						"$ref": "#/components/schemas/DeniedGetUserPageResponseReason"
					}
				},
				"required": [
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DisplayedPost": {
				"properties": {
					"imageUrl": {
						"type": "string"
					},
					"creatorUsername": {
						"type": "string"
					},
					"creationTimestamp": {
						"type": "number",
						"format": "double"
					},
					"caption": {
						"type": "string"
					},
					"likes": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"count"
						],
						"type": "object"
					},
					"comments": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"count"
						],
						"type": "object"
					},
					"shares": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"count"
						],
						"type": "object"
					}
				},
				"required": [
					"imageUrl",
					"creatorUsername",
					"creationTimestamp",
					"caption",
					"likes",
					"comments",
					"shares"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DisplayedShopItems": {
				"properties": {
					"title": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"countSold": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"price",
					"description",
					"countSold"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessfulGetUserPageResponse": {
				"properties": {
					"username": {
						"type": "string"
					},
					"followers": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"count"
						],
						"type": "object"
					},
					"subscribers": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"count"
						],
						"type": "object"
					},
					"follows": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"count"
						],
						"type": "object"
					},
					"bio": {
						"type": "string"
					},
					"posts": {
						"items": {
							"$ref": "#/components/schemas/DisplayedPost"
						},
						"type": "array"
					},
					"shopItems": {
						"items": {
							"$ref": "#/components/schemas/DisplayedShopItems"
						},
						"type": "array"
					}
				},
				"required": [
					"username",
					"followers",
					"subscribers",
					"follows",
					"bio",
					"posts",
					"shopItems"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HTTPResponse_DeniedGetUserPageResponse.SuccessfulGetUserPageResponse_": {
				"properties": {
					"error": {
						"$ref": "#/components/schemas/DeniedGetUserPageResponse"
					},
					"success": {
						"$ref": "#/components/schemas/SuccessfulGetUserPageResponse"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GetUserPageParams": {
				"properties": {
					"username": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SecuredHTTPRequest_GetUserPageParams_": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetUserPageParams"
					}
				},
				"required": [
					"accessToken",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playhouse-backend-2",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPResponse_FailedAuthResponse.SuccessfulAuthResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserParams"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPResponse_FailedAuthResponse.SuccessfulAuthResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginUserParams"
							}
						}
					}
				}
			}
		},
		"/auth/resetPassword": {
			"post": {
				"operationId": "RequestPasswordReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPResponse_DeniedPasswordResetResponse.SuccessfulPasswordResetResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestPasswordResetParams"
							}
						}
					}
				}
			}
		},
		"/post/create": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPResponse_FailedToCreatePostResponse.SuccessfulPostCreationResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SecuredHTTPRequest_CreatePostParams_"
							}
						}
					}
				}
			}
		},
		"/user/SetSettings": {
			"post": {
				"operationId": "SetUserSettings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPResponse_FailedToUpdateUserSettingsResponse.SuccessfulUpdateToUserSettingsResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SecuredHTTPRequest_SetUserSettingsParams_"
							}
						}
					}
				}
			}
		},
		"/user/GetPosts": {
			"post": {
				"operationId": "GetPostsPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPResponse_DeniedGetUserPageResponse.SuccessfulGetUserPageResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SecuredHTTPRequest_GetUserPageParams_"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}